requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# HTTP Requests
requests==2.31.0
httpx==0.25.1

# Environment Variables
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Technical Indicators
ta-lib==0.4.28
pandas-ta==0.3.14b0

# Caching
cachetools==5.3.2

# Logging
loguru==0.7.2

# Date/Time
python-dateutil==2.8.2

# Scientific Computing (for advanced indicators)
scipy==1.11.4

# Async support
aiofiles==23.2.1
